knt m3e

+++ src/HomePage/Home.jsx
@@ -2,9 +2,13 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import {useState} from 'react';
 import axios from "axios";
 
-export default function Home({tableData,setTableData,handleLoggin,setEmail,setPass,setCPass}) {
+export default function Home({tableData,setTableData,handleLoggin,Email,setEmail,setPass,setCPass,role,setisSaving,handleRegistring}) {
 
+    console.log('UserRole in Home:', role);
+    console.log('UserEamil in Home:', Email);
+    
 
+    
     const [showInputFields, setShowInputFields] = useState(false);
 
     const handlelogout = () => {
@@ -63,8 +67,8 @@ export default function Home({tableData,setTableData,handleLoggin,setEmail,setPa
                     picture: formData.picture,
                     allowed: formData.canDelete
                 });
-
-                setTableData(previousData => [...previousData, { id: tableData.length+1, ...formData }]);
+                setisSaving(true)
+                //setTableData(previousData => [...previousData, { id: tableData.length+1, ...formData }]);
 
         }
          else {
@@ -73,23 +77,23 @@ export default function Home({tableData,setTableData,handleLoggin,setEmail,setPa
     };
 
 
-
-
-
-
-
-
     const handleDeleteButtonClick = (id) => {
-        // Make DELETE request to the server
-        axios.delete(`http://localhost:5174/api/delete/${id}`)
+        // Check if the user has the necessary role before making the delete request
+        if (role === 'admin') {
+          // Make DELETE request to the server
+          axios.delete(`http://localhost:5174/api/delete/${id}`)
             .then(() => {
-                // If deletion on the server is successful, update the state
-                setTableData((prevTableData) => prevTableData.filter((data) => data.id !== id));
+              // If deletion on the server is successful, update the state
+              setTableData((prevTableData) => prevTableData.filter((data) => data.id !== id));
             })
             .catch((error) => {
-                console.error('Error deleting record:', error);
+              console.error('Error deleting record:', error);
             });
-    };
+        } else {
+          // User does not have the necessary privileges
+          alert('You do not have permission to delete employees.');
+        }
+      };
 
 
 
@@ -224,6 +228,16 @@ export default function Home({tableData,setTableData,handleLoggin,setEmail,setPa
             >
                 Log Out
             </button>
+
+            <button
+                type="button"
+                className="btn btn-primary mt-3"
+                onClick={handleRegistring}
+                style={{display: role==="admin"? "block":"none"}}
+            >
+                Add User
+            </button>
+
         </>
     );
 }
